feat: move function for old site in new file and add file for parsing functions
